added: prototype color lighten/darken script
ideally the script that will (in the future) build the matching gradience and
kvantum themes will be passed all the correspinding colors, however it needs to
be able to handle the case in which a reduced color palette is passed, and thus
requires to calculate some additional shades.

currently this script is using the shell built in math, that rounds divisions
with a bias towards ceiling, while python rounds towards floor, will have to
use awk for math to round towards floor.