add kvantum-theme builder
initial work on the script to build the kvlibadwaita theme properly instead
of just templating with pywal.

so far the values from any gradience json theme are loaded on a dict, the
reference values "@key-name" are replaced by the value they point to, still
need to convert the rgba values to hex strings.